Made small data type optimizations.

+++ graphics_hw_cpp/include/rasterizer.hpp
@@ -9,12 +9,12 @@
 
 class Fetch {
 private:
-	std::queue<std::array<int, 3>> indices;
+	std::queue<std::array<int16_t, 3>> indices;
 	Clock* clk;
 public:
 
-	std::array<std::array<int, 3>, 2> forward(Status* FE_BB, std::array<std::array<int, 3>, 2> batch);
-	void comb(Status* FE_BB, std::array<int, 3> tri);
+	std::array<std::array<int16_t, 3>, 2> forward(Status* FE_BB, std::array<std::array<int16_t, 3>, 2> batch);
+	void comb(Status* FE_BB, std::array<int16_t, 3> tri);
 
 	Fetch(Clock* clk);
 
@@ -22,13 +22,13 @@ public:
 
 class BoundingBox {
 private:
-	std::queue<std::array<int, 3>> indices;
+	std::queue<std::array<int16_t, 3>> indices;
 	std::queue<std::array<std::array<f16Vector2, 2>, 2>> bounding_box;
 	Clock* clk;
 public:
 
-	std::array<std::array<int, 3>, 2> forward(Status* BB_DP, Status* FE_BB, std::array<std::array<int, 3>, 2> batch);
-	void comb(Status* FE_BB, std::array<std::array<int, 3>, 2> tris, VectorTable* table);
+	std::array<std::array<int16_t, 3>, 2> forward(Status* BB_DP, Status* FE_BB, std::array<std::array<int16_t, 3>, 2> batch);
+	void comb(Status* FE_BB, std::array<std::array<int16_t, 3>, 2> tris, VectorTable* table);
 	std::array<std::array<f16Vector2, 2>, 2> getBB();
 
 	BoundingBox(Clock* clk);
